test(e2e): split 814-line core-worktree-resource-lifecycle.spec.ts into 3 thematic specs
- Extract 8 shared helper functions to e2e/helpers/resource-lifecycle.ts
- Split lifecycle tests (create, delete, teardown) into core-worktree-resource-lifecycle.spec.ts (5 tests)
- Split resource ops tests (status, pause, resume, provision) into core-worktree-resource-ops.spec.ts (6 tests)
- Split command substitution tests (connect, env vars, templates) into core-worktree-resource-substitution.spec.ts (3 tests)
- Update platform spec to import commandArg and nodeScriptCommand from shared helper
- Each spec is independently runnable with its own beforeAll/afterAll, improving CI shard distribution
